fix debian

+++ scripts/regex_extract.sh
@@ -1,3 +1,12 @@
 mkdir -p package
 grep -oP '\/usr[^\>]+' $1 |  xargs -I '{}' cp '{}' package
-cp /usr/lib64/ld-linux-x86-64.so.2 package
+
+if [ -f /usr/lib64/ld-linux-x86-64.so.2 ]
+then
+    cp /usr/lib64/ld-linux-x86-64.so.2 package
+fi
+
+if [ -f /lib64/ld-linux-x86-64.so.2]
+then
+    cp /lib64/ld-linux-x86-64.so.2 package
+fi
